v1.2.0 -- Added Config Passwords
This update will require a fresh uninstall and reinstall if its already installed, this is to accommodate for the database structure changes.  In this update, configuration passwords were added to help protect from any unwanted changes.  This is obviously optional and can be added or removed at any time in the new 'Access' tab withing the config menu.